+++ examples/radio_buttons/src/main.rs
@@ -1,44 +1,44 @@
 //! A simple radio button example using the built-in RadioButton widget.
 //!
-//! Note that in this example we override the default widget spec with some spacing adjustments. Check out
-//! the `radio_button` scene in `assets/main.cob.json`.
+//! Note that in this example we build the button scenes from scratch in `assets/main.cob`. Other examples may use
+//! built-in default styling for radio buttons.
 
 use bevy::prelude::*;
 use bevy::window::WindowTheme;
-use bevy_cobweb_ui::builtin::widgets::radio_button::{RadioButtonBuilder, RadioButtonManager};
 use bevy_cobweb_ui::prelude::*;
 
 //-------------------------------------------------------------------------------------------------------------------
 
 fn build_ui(mut c: Commands, mut s: ResMut<SceneLoader>)
 {
-    let file = &SceneFile::new("examples.radio_buttons");
     static OPTIONS: [&'static str; 3] = ["A", "B", "C"];
 
-    let scene = file + "scene";
+    let scene = ("main.cob", "scene");
     c.ui_root().load_scene_and_edit(scene, &mut s, |l| {
         // Get the display text's entity.
         let display_text = l.get("display::text").id();
 
         // Insert radio buttons.
         l.edit("radio_frame", |l| {
-            let n = l.deref_mut();
-            let manager_entity = RadioButtonManager::insert(n);
-
             for (i, option) in OPTIONS.iter().enumerate() {
-                // Add radio button.
-                let entity = RadioButtonBuilder::custom_with_text(file + "radio_button", *option)
-                    .with_indicator()
-                    .build(manager_entity, n)
-                    .on_select(move |mut e: TextEditor| {
+                l.load_scene_and_edit(("main.cob", "button"), |l| {
+                    // Set display text when a button is selected.
+                    l.on_select(move |mut e: TextEditor| {
                         write_text!(e, display_text, "Selected: {}", option);
-                    })
-                    .id();
+                    });
 
-                // Select the first option.
-                if i == 0 {
-                    n.react().entity_event(entity, Select);
-                }
+                    l.get("text").update(|id| {
+                        move |mut e: TextEditor| {
+                            write_text!(e, id, "{}", option);
+                        }
+                    });
+
+                    // Select the first option.
+                    if i == 0 {
+                        let entity = l.id();
+                        l.react().entity_event(entity, Select);
+                    }
+                });
             }
         });
     });
@@ -61,7 +61,7 @@ fn main()
             ..default()
         }))
         .add_plugins(CobwebUiPlugin)
-        .load("main.cob.json")
+        .load("main.cob")
         .add_systems(PreStartup, setup)
         .add_systems(OnEnter(LoadState::Done), build_ui)
         .run();